postman

+++ ewm-main-service/src/main/java/ru/practicum/events/dto/EventShortDto.java
@@ -18,5 +18,5 @@ public class EventShortDto {
     private UserShortDto initiator;
     private boolean paid;
     private String title;
-    private Long views;
+    private int views;
 }
